Added ReplicaId Counter in PM and is passed by rpc call

+++ PuppetMasterCLI/PuppetMaster.cs
@@ -25,12 +25,15 @@ namespace PuppetMasterCLI
         private bool debug = false;
         private Process schedulerProcess;
 
+        private int replicaIdCounter;
+
         public PuppetMaster(List<string> pcsList)
         {
             foreach (string pcsURL in pcsList)
             {
                 var pcsHost = pcsURL.Split("//")[1].Split(":")[0];
                 pcsManagers.Add(pcsHost, new PCSManager(pcsURL));
+                replicaIdCounter = 0;
             }
         }
 
@@ -70,7 +73,7 @@ namespace PuppetMasterCLI
         public void CreateStorage(string serverId, string url, int gossipDelay)
         {
             var pcsHost = url.Split("//")[1].Split(":")[0];
-            pcsManagers[pcsHost].createStorageNode(serverId, url, gossipDelay);
+            pcsManagers[pcsHost].createStorageNode(serverId, url, gossipDelay, replicaIdCounter);
 
             StorageNodeStruct node;
             node.serverId = serverId;
@@ -80,6 +83,7 @@ namespace PuppetMasterCLI
 
             storageNodes.Add(node);
 
+            replicaIdCounter++;
         }
         public void CreateWorker(string serverId, string url, int gossipDelay)
         {
@@ -252,7 +256,8 @@ namespace PuppetMasterCLI
             {
                 pcs.Value.exit();
             }
-            schedulerProcess.Kill();
+            if (schedulerProcess != null)
+                schedulerProcess.Kill();
             WaitForSchedulerProcess();
             ls.ShutDown();
         }
